@@ -83,14 +83,15 @@
     "postcss-import": "^8.1.2",
     "postcss-loader": "^0.9.1",
     "postcss-modules-local-by-default": "^1.1.1",
-    "react": "15.3.0",
-    "react-addons-test-utils": "15.3.0",
-    "react-dom": "15.3.0",
+    "react": "15.4.1",
+    "react-addons-test-utils": "~15.1.0",
+    "react-dom": "~15.4.1",
     "react-google-maps": "^6.0.1",
     "react-hot-loader": "^1.3.0",
     "react-redux": "^4.4.2",
     "react-router": "^2.4.1",
     "react-router-redux": "^4.0.0",
+    "react-tap-event-plugin": "^2.0.1",
     "redux": "^3.3.1",
     "redux-debounced": "^0.3.0",
     "redux-form": "^5.3.2",
