@@ -2,11 +2,13 @@ import React from 'react';
 import { Route } from 'react-router';
 import App from '../app/app.js';
 import Stores from '../containers/stores/stores';
-// import Nav from '../containers/navigation/navigation';
+import Nav from '../containers/navigation/navigation';
 
 export default (
   <div>
-	<Route path="/" component={ App }/>
-	<Route path="/stores" component={ Stores }/>
+  	<Route path="" component={ Nav }>
+    	<Route path="/" component={ App }/>
+    	<Route path="/stores" component={ Stores }/>
+    </Route>
   </div>
 );
